website/requirements.txt
bcrypt==3.1.4
cffi==1.11.5
config==0.4.0
Django==1.11.16
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
psycopg2==2.7.5
pycparser==2.19
PyJWT==1.6.4
pytz==2018.7
six==1.11.0
